sperate migration and schema mangers
